+++ backend/controllers/outcomes.js
@@ -1,10 +1,18 @@
 const outcomesRouter = require('express').Router()
 const Outcome = require('../models/outcome')
 const Game = require('../models/game')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 
 outcomesRouter.get('/', async (request, response) => {
-//await Outcome.deleteMany({})
   const outcomes = await Outcome.find({})
     .populate({
       path: 'game',
@@ -26,7 +34,20 @@ outcomesRouter.get('/', async (request, response) => {
 outcomesRouter.post('/', async (request, response) => {
   const body = request.body
 
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(400).end()
+  }
+
   const game = await Game.findById(body.game)
+
+
+  const date = new Date(game.date)
+  const now = new Date()
+
+  if (date > now) {
+    return response.status(400).json({ error: 'Outcome can not be added for future games' })
+  }
 
   const outcome = new Outcome({
     goals_home: Number(body.goals_home),
@@ -51,6 +72,10 @@ outcomesRouter.get('/:id', async (request, response) => {
 })
 
 outcomesRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(400).end()
+  }
   await Outcome.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
@@ -58,6 +83,10 @@ outcomesRouter.delete('/:id', async (request, response) => {
 
 outcomesRouter.put('/:id', (request, response, next) => {
   const { goals_home, goals_visitor, game } = request.body
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(400).end()
+  }
 
   Outcome.findByIdAndUpdate(
     request.params.id,
